+++ src/Routes/Tables/Jobs/JobTypes.js
@@ -3,7 +3,7 @@ import { FlexBox } from 'components/common';
 import PropTypes from 'prop-types';
 import React from 'react';
 import styled from 'styled-components';
-import { COLOR_PIPELINE_TYPES, COLOR_PIPELINE_TYPES_TEMPLATE } from 'styles';
+import { COLOR_PIPELINE_TYPES, Theme } from 'styles';
 import { toUpperCaseFirstLetter as toUpper } from 'utils';
 
 const Overflow = styled(FlexBox.Auto)`
@@ -14,33 +14,33 @@ const Overflow = styled(FlexBox.Auto)`
 const CapitalizedTag = styled(Tag)`
   text-transform: capitalize;
   border: 1px solid
-    ${props => (props.theme.isDarkMode ? props.$borderType : 'inherit')};
+    ${props =>
+      Theme.Styles.CapitalizedTag?.borderType[props.$borderType] || 'inherit'};
 `;
 
-const JobTypes = ({ types, fullName }) =>
-  types && (
-    <Overflow justify="start" gutter={0}>
-      {types !== undefined &&
-        types.map(type =>
-          fullName ? (
+const JobTypes = ({ types, fullName }) => (
+  <Overflow justify="start" gutter={0}>
+    {types !== undefined &&
+      types.map(type =>
+        fullName ? (
+          <CapitalizedTag
+            key={type}
+            color={Theme.Styles.isTagFill ? COLOR_PIPELINE_TYPES[type] : ''}
+            $borderType={type}>
+            {type}
+          </CapitalizedTag>
+        ) : (
+          <Tooltip key={type} placement="top" title={toUpper(type)}>
             <CapitalizedTag
-              key={type}
-              color={COLOR_PIPELINE_TYPES_TEMPLATE[type]}
-              $borderType={COLOR_PIPELINE_TYPES[type]}>
-              {type}
+              color={Theme.Styles.isTagFill ? COLOR_PIPELINE_TYPES[type] : ''}
+              $borderType={type}>
+              {type.slice(0, 2)}
             </CapitalizedTag>
-          ) : (
-            <Tooltip key={type} placement="top" title={toUpper(type)}>
-              <CapitalizedTag
-                color={COLOR_PIPELINE_TYPES_TEMPLATE[type]}
-                $borderType={COLOR_PIPELINE_TYPES[type]}>
-                {type.slice(0, 2)}
-              </CapitalizedTag>
-            </Tooltip>
-          )
-        )}
-    </Overflow>
-  );
+          </Tooltip>
+        )
+      )}
+  </Overflow>
+);
 
 JobTypes.propTypes = {
   types: PropTypes.arrayOf(PropTypes.string),
