+++ src/components/index.js
@@ -1,3 +1,2 @@
 export { default as ErrorBoundary } from './ErrorBoundary';
 export { default as Table } from './Table';
-export { default as GlobalStyle } from './GlobalThemeStyle';
