+++ src/const/themesNames.js
@@ -1,4 +1,4 @@
 export const THEMES_NAMES = {
-  Light: 'light',
-  Dark: 'dark',
+  LIGHT: 'light',
+  DARK: 'dark',
 };
