+++ src/hooks/useSiteDarkMode.react.js
@@ -1,5 +1,6 @@
 import { useCallback, useState, useRef } from 'react';
 import { THEMES_NAMES } from 'const';
+
 import { createStore } from 'reusable';
 import { useHistory } from 'react-router';
 
@@ -10,40 +11,43 @@ const useSiteDarkMode = () => {
   const systemTheme = () =>
     window.matchMedia &&
     window.matchMedia('(prefers-color-scheme: dark)').matches
-      ? THEMES_NAMES.Dark
-      : THEMES_NAMES.Light;
+      ? THEMES_NAMES.DARK
+      : THEMES_NAMES.LIGHT;
 
   // get state theme from user local storage
-  const getTheme = () => localStorage.getItem('theme') || systemTheme();
+  const getTheme = () => {
+    const themeName = localStorage.getItem('theme') || systemTheme();
+    return themeName !== null && typeof themeName !== 'undefined'
+      ? themeName
+      : 'light';
+  };
 
   // set if is state dark mode
   const [themeName, setThemeName] = useState(getTheme());
-  const [isDarkMode, setDarkMode] = useState(themeName === THEMES_NAMES.Dark);
   const refContainer = useRef(0);
 
   const setTheme = useCallback(
     nameTheme => {
       localStorage.setItem('theme', nameTheme);
-      setThemeName(nameTheme);
+      setThemeName(themeName);
       // reloads the page after change theme.
       if (refContainer.current > 0) history.go(0);
       else {
-        setDarkMode(nameTheme === THEMES_NAMES.Dark);
         refContainer.current = 1;
       }
     },
-    [history]
+    [history, themeName]
   );
 
   // toggle switch between dark and light
-  const toggleTheme = () =>
+  const toggleTheme = () => {
+    const cTheme = getTheme();
     setTheme(
-      getTheme() === THEMES_NAMES.Dark ? THEMES_NAMES.Light : THEMES_NAMES.Dark
+      cTheme === THEMES_NAMES.DARK ? THEMES_NAMES.LIGHT : THEMES_NAMES.DARK
     );
+  };
 
   return {
-    isDarkMode,
-    setDarkMode,
     toggleTheme,
     setTheme,
     themeName,
