+++ src/styles/BottomPosition.styles.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { COLOR_LAYOUT } from './colors';
+import { COLOR_LAYOUT } from 'styles/colors';
 
 const BottomPosition = styled.div`
   position: absolute;
