+++ src/styles/index.js
@@ -7,9 +7,7 @@ export {
   COLOR_LAYOUT,
   COLOR_PIPELINE_STATUS,
   COLOR_PIPELINE_TYPES,
-  COLOR_PIPELINE_TYPES_TEMPLATE,
   COLOR_PRIORITY,
-  COLOR_PRIORITY_TEMPLATE,
   COLOR_SERVICE,
   COLOR_TASK_STATUS,
 } from './colors';
