+++ src/styles/themes/dark/DarkThemeStyle.styles.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import './dark-mode-style.css';
+
+const DarkThemeStyle = () => <> </>;
+
+export default DarkThemeStyle;
