+++ src/styles/themes/light/LightThemeStyle.styles.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import './light-mode-style.css';
+
+const LightThemeStyle = () => <> </>;
+
+export default LightThemeStyle;
