+++ src/styles/themes/light/light-mode-style.css
@@ -1,0 +1,3 @@
+.ant-layout {
+  background-color: white;
+}
