--- src/styles/themes/themes.js
@@ -1,2 +1,0 @@
-export { default as Dark } from './dark';
-export { default as Light } from './light';
